test(e2e): mobile-viewport onboarding→home lane + shared fixtures (#9525)
Pixel-7 touch-viewport counterpart of onboarding-to-home.spec.ts driving the
same keyless first-run → home → swipe-springboard flow; extracts the shared
fixtures/route-mocks/assertions into onboarding-to-home.shared.ts (both specs
use it) and wires the mobile lane into scenario-pr.yml CI browser coverage.
